added job pooling

+++ mj_automation.py
@@ -1,6 +1,7 @@
 import asyncio
 import discord
 from discord.ext import commands
+from discord.errors import DiscordServerError
 from dotenv import load_dotenv
 import os
 import requests
@@ -22,6 +23,8 @@ class MjAutomator:
         self.done_prompt_file = open(config.DONE_PROMPT_FILE, 'a')
         self.auto_run = auto_run
         self.log_started = False
+        self.message_counter = 0
+        self.job_counter = 0
 
         self.intents = discord.Intents.all()
         self.intents.message_content = True
@@ -31,6 +34,7 @@ class MjAutomator:
         self.downloader = self.Downloader(self) if config.ENABLE_DOWNLOAD else None  # the downloader class
         self.prompter = self.Prompter(self) if config.ENABLE_PROMPTING else None  # the prompter class
         self.mj_imagine, self.mj_upscale, self.mj_upscale_max = self.discord_setup()  # bringing the inner methods to the main class
+        self.job_manager = self.JobManager(self)  # the job manager class, handling pool of all midjourney commmands
 
         if self.auto_run:
             self.start_bot()
@@ -39,13 +43,13 @@ class MjAutomator:
         self.client.run(token)
 
     async def quit_bot(self):
-        asyncio.run(self.client.close())
+        asyncio.create_task(self.client.close())
 
     def discord_setup(self):
         @self.client.event
         async def on_ready():
             # flush the log file if it's not persistent
-            if config.LOG_ENABLED and not config.LOG_PERSISTENT:
+            if config.LOG_ENABLED and not config.LOG_PERSISTENT and os.path.exists(config.LOG_FILE):
                 os.remove(config.LOG_FILE)
 
             # Get a reference to the guild (server) and channel
@@ -53,9 +57,9 @@ class MjAutomator:
             self.channel = self.guild.get_channel(config.CHANNEL_ID)
 
             print(f"Logged in as {self.client.user}")
-            print(f"Downloader status: {'enabled' if self.downloader else 'disabled'}")
-            print(f"Prompter status: {'enabled' if self.prompter else 'disabled'}")
+            print(f"Downloader: {'enabled' if self.downloader else 'disabled'}, prompter: {'enabled' if self.prompter else 'disabled'}")
 
+            asyncio.create_task(self.job_manager.process_jobs())
             await self.channel.send("Bot ready!")
             if self.prompter and self.auto_run:
                 asyncio.create_task(self.prompter.start_prompting())
@@ -194,27 +198,12 @@ class MjAutomator:
             await self.main.channel.send('Starting prompting...')
 
             prompts = self.main.prompt_file.readlines()
-            prompt_counter = 0
-            number_of_prompts = len(prompts)
 
-            while prompt_counter < number_of_prompts:
-                print(f'\rprompt {prompt_counter + 1}/{number_of_prompts}', end='', flush=True)
-                prompt = prompts[prompt_counter].strip()
+            for prompt in prompts:
+                prompt = prompt.strip()
                 if prompt != "":
-                    await self.send_prompt(prompt)
-                prompt_counter += 1
-                if prompt_counter % 10 == 0:
-                    await asyncio.sleep(90)  # sleep for 90 seconds every 10 prompts
-                await asyncio.sleep(30)
-            print('\nPrompt batch finished.')
-            if config.QUIT_ON_COMPLETION:
-                await asyncio.sleep(120)  # give time to finish downloading jobs
-                await self.client.close()
-                print('Bot closed.')
-            else:
-                # enter infinite loop to keep the thread running
-                while True:
-                    await asyncio.sleep(60)
+                    await self.main.job_manager.add_job((self.main.prompter.send_prompt, prompt))
+            print('Prompt batch import finished. Created jobs:', await self.main.job_manager.get_job_count())
 
         async def send_prompt(self, prompt):
             # Check if there are any messages in the channel
@@ -255,19 +244,61 @@ class MjAutomator:
                 # choose upscaler based on message contents, but ignore already max_upscaled images
                 if any(s in lowered_message_content for s in config.UPSCALE_TAGS) and \
                         config.UPSCALE_MAX_TAG not in lowered_message_content and config.ENABLE_UPSCALE_MAX:
-                    await self.max_upscale(ctx, message_id, message_hash)
+                    # await self.max_upscale(ctx, message_id, message_hash)
+                    # adding to job queue
+                    await self.main.job_manager.add_job((self.main.prompter.max_upscale, ctx, message_id, message_hash))
                 else:
                     # if it's already a max upscaled image, don't upscale it again as it causes errors
                     if config.UPSCALE_MAX_TAG not in lowered_message_content:
-                        await self.default_upscale(ctx, message_id, message_hash)
+                        # await self.default_upscale(ctx, message_id, message_hash)
+                        # adding to job queue
+                        await self.main.job_manager.add_job((self.main.prompter.default_upscale, ctx, message_id, message_hash))
 
             except Exception as e:
                 print(f"An error occurred: {e}")
 
         async def default_upscale(self, ctx, message_id, message_hash):
-            for i in range(1, 2):
+            for i in range(1, 5):
                 await self.main.mj_upscale(ctx, i, message_id, message_hash)
                 await asyncio.sleep(5)
 
         async def max_upscale(self, ctx, message_id, message_hash):
             await self.main.mj_upscale_max(ctx, message_id, message_hash)
+
+    class JobManager:
+        def __init__(self, main):
+            self.main = main
+            self.client = self.main.client
+            self.jobs = []
+
+        async def add_job(self, job):
+            self.jobs.append(job)
+
+        async def remove_job(self):
+            self.jobs.pop(0)
+
+        def get_next_job(self):
+            return self.jobs[0] if self.jobs else None
+
+        async def get_job_count(self):
+            return len(self.jobs)
+
+        async def do_job(self, job):
+            job_function, *args, kwargs = job
+            await job_function(*args, kwargs)
+            await self.remove_job()
+
+        async def process_jobs(self):
+            print("Starting job processing...")
+
+            while True:
+                job = self.get_next_job()
+                if job is not None:
+                    try:
+                        await self.do_job(job)
+                    except DiscordServerError as e:
+                        print(f"DiscordServerError: {e}")
+                        await asyncio.sleep(60)
+
+                await asyncio.sleep(30)  # sleep for 30 seconds between jobs
+
